Refactored PR2101 and checked with logs
Certain problems were showing with PR2101, like not showing whereami ..

So refactored - moved to a enum for docked/landed/hyperspace so we can't get into strange conditions with the previous bools

Removed Nolocation - the enum copes with it.

Checked bodyapproach is okay

added died and loadout handling
Fixed touchdown - you can have a npc controlled touchdown

checked it against logs.
